Rename math_utils for export
