+++ langchain/src/util/math.ts
